Add code to handle the ccdscan case as a normal rsMap.  In this case the SPEC file has #CCD lines describing the ccd.  #CCD is also in the xpcsscan but in this case one of the lines is like
#CCD ccdscan NUMBER 

where number is the scan number.  At this point the files that have been
produced on the beamline have the #CCD lines before the #S line
indicating the start of the scan.  Have reached an agreement with Eric
Dufresne that they will move this after the #S since spec2nexus puts
everything before the #S line in the scan before which creates more
complicated logic.  ccdscans are actually ascan, a2scan, etc. but have
these #CCD lines, most specifically the one described above.  This type
of scan defines the original location  of the data files, the # of dark
images and # of repeats.  So far, this is handling moving the file to a
new location and subtractign the Dark images.  So far, the darks are
averaged but ignore the first dark,  Will need to go back and add a
number of initial darks to skip.I do not have an example of repeat
images so I am not yet handling this case.